pyproject.toml
[build-system]
requires = [ "setuptools" ]
build-backend = "setuptools.build_meta"

[project]
name = "axyn"
version = "8.16.2"
authors = [
  { name = "Daniel Thwaites", email = "danth@danth.me" },
]
description = "A discord chatbot using traditional algorithms"
readme = "README.md"
license = "AGPL-3.0-or-later"
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.13",
]
keywords = [ "Discord", "bot", "chatbot" ]
requires-python = "~= 3.13"
dependencies = [
  "alembic >=1,<2",
  "discord.py >=2,<3",
  "fastembed >=0.7,<1",
  "ngt >=2,<3",
  "opentelemetry-api ~= 1.34",
  "opentelemetry-exporter-otlp-proto-http ~= 1.34",
  "opentelemetry-instrumentation-sqlalchemy ~= 0.55b0",
  "opentelemetry-sdk ~= 1.34",
  "sqlalchemy[aiosqlite] >=2,<3",
  "aiosqlite <0.22", # Causes tests to hang.
]

[project.scripts]
axyn = "axyn.__main__:launch"

[project.optional-dependencies]
build = [ "build >=1,<2" ]
test = [
  "coverage >=7,<8",
  "pytest >=8,<9",
  "pytest-asyncio >=1,<2",
]

[project.urls]
changelog = "https://github.com/danth/axyn/releases"
homepage = "https://github.com/danth/axyn"
issues = "https://github.com/danth/axyn/issues"
source = "https://github.com/danth/axyn"

[tool.coverage.run]
branch = true
source = [ "axyn" ]

[tool.pyright]
typeCheckingMode = "strict"
reportMissingTypeStubs = false
pythonVersion = "3.13" # Keep in sync with setup.py

[tool.pytest.ini_options]
addopts = [ "--import-mode=importlib" ]
asyncio_mode = "auto"

[tool.semantic_release]
commit_parser = "conventional"
commit_message = "build: release v{version}"
version_toml = [ "pyproject.toml:project.version" ]
version_variables = [ "flake.nix:axynVersion" ]

[tool.setuptools.packages.find]
include = [ "axyn*" ]

